CloudComputing/AmazonCloudServices/pythonProject/requirements.txt
grpcio
boto3
protobuf
googleapis-common-protos